% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_private.R
\name{plot.VTPRIOR}
\alias{plot.VTPRIOR}
\title{Plot prior distributions}
\usage{
\method{plot}{VTPRIOR}(x, model = 0, n = 10000, draw.levels = NULL,
  adjust = 1.2, ...)
}
\arguments{
\item{x}{A class \code{VTPRIOR} list generated by \code{\link{vtPriorPar}}}

\item{model}{0: parametric}

\item{n}{?}

\item{draw.levels}{?}

\item{adjust}{?}

\item{...}{?}
}
\description{
Plot densities of DLT and response rate based on prior knowledge
}
