% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_private.R
\name{summary.VTPRIOR}
\alias{summary.VTPRIOR}
\title{Summary of Priors}
\usage{
\method{summary}{VTPRIOR}(object, qs = c(0.025, 0.5, 0.975), model = 0,
  n = 10000, ...)
}
\arguments{
\item{object}{?}

\item{qs}{?}

\item{model}{0:PARA; 1:PARA+; 2:PARA SINGLEALPHA; 3:PARA+ SINGLEALPHA}

\item{n}{?}

\item{...}{?}
}
\description{
Summary of DLT and response rate based on prior knowledge
Doesn't work
}
