% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_simu.R
\name{summary.VTTRUEPS}
\alias{summary.VTTRUEPS}
\title{Print true probabilities}
\usage{
\method{summary}{VTTRUEPS}(object, digits = 2, ...)
}
\arguments{
\item{object}{A class \code{VTTRUEPS} matrix generated by \code{\link{vtScenario}}}

\item{digits}{Digits for print}

\item{...}{Reserved parameters}
}
\value{
A table showing the summary of the \code{VTTRUEPS} object. The first
    four columns are individual probability, fifth and sixth are probability
    for toxicity and resistance, and seventh is rho, the correlation.
}
\description{
Print the true probabilities, with probabilities of toxicity and resistance,
     and \eqn{\rho}.
}
\examples{
rst.sce <- vtScenario(tox = c(0.05, 0.05, 0.08),
                      res = c(0.2, 0.3, 0.5),
                      rho = 1)
summary(rst.sce)

}
