% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_simu.R
\name{summary2.VTTRUEPS}
\alias{summary2.VTTRUEPS}
\title{Print true probabilities in latex format}
\usage{
\method{summary2}{VTTRUEPS}(x, rp2d = -1, digits = 2, ...)
}
\arguments{
\item{x}{A class \code{VTTRUEPS} matrix generated by \code{\link{vtScenario}}}

\item{rp2d}{Columns to be in bold font}

\item{digits}{Digits for print}

\item{...}{Reserved parameters}
}
\value{
A summary of the true probabilities in latex format.
}
\description{
Print the true probabilities, with probabilities of toxicity and resistance,
     and \eqn{\rho}, in latex format
}
\examples{

rst.sce <- vtScenario(tox = c(0.05, 0.05, 0.08),
                      res = c(0.2, 0.3, 0.5),
                      rho = 1)
ltx.ps <- summary2(rst.sce)


}
