% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_shiny.R
\name{vtShiny}
\alias{vtShiny}
\title{Run Web-Based \code{visit} application}
\usage{
vtShiny()
}
\description{
Call Shiny to run \code{visit} as a web-based application.
}
\details{
A web browser will be brought up for users to access the GUI of \code{\link{visit}}.
}
\examples{
if(interactive()){
vtShiny()}

}
